+++ src/utils/constants.rs
@@ -8,5 +8,5 @@ pub const REGISTER_FILENAME: &str = "Register.toml";
 pub const CONFIG_FILENAME: &str = "Config.toml";
 
 pub const DEFAULT_METADATA_REPOSITORY_NAME: &str = "core";
-pub const DEFAULT_METADATA_REPOSITORY_PATH: &str = "https://raw.githubusercontent.com/pack-it/core/main/";
+pub const DEFAULT_METADATA_REPOSITORY_URL: &str = "https://raw.githubusercontent.com/pack-it/core/main/";
 pub const DEFAULT_METADATA_REPOSITORY_PROVIDER: &str = WEB_METADATA_PROVIDER_ID;