+++ src/verifier/repairer.rs
@@ -151,7 +151,7 @@ impl Repairer {
         let mut seen_repositories = HashMap::new();
         for package in register.iterate_all() {
             let repository = Repository {
-                path: package.metadata_repository_url.clone(),
+                url: package.metadata_repository_url.clone(),
                 provider: package.metadata_repository_provider.clone(),
                 prebuilds_url: package.prebuilds_repository_url.clone(),
                 prebuilds_provider: package.prebuilds_repository_provider.clone(),